mobile: fix lints - replace withOpacity() with withValues(alpha:) and await ref.refresh().future

+++ mobile/destiny_decoder_app/lib/features/decode/presentation/decode_result_page.dart
@@ -433,7 +433,7 @@ class _DecodeResultPageState extends ConsumerState<DecodeResultPage>
                     FloatingActionButton.extended(
                       heroTag: 'daily_insights',
                       onPressed: () {
-                        final lifeSealNumber = lifeSeal.number as int;
+                        final lifeSealNumber = lifeSeal.number;
                         final dayOfBirth = DateTime.parse(result.input.dateOfBirth).day;
                         Navigator.of(context).push(
                           MaterialPageRoute(
